requirements-dev.txt
# Development and testing dependencies
-r requirements.txt

pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0
